add reports workflow
this workflow allows bulk export/import of report. the user will be presented with a list of all reports in their PT instance. Reports will not be sorted by client, if client matters they should be using the client/repotrs workflow.

On import, all selected PTRAC files will be imported into a single selected client.

moved functions to the utils file that are called in both client/reports and reports workflow