add option to Clients and Reports workflow to update existing clients
if option is selected the script will look at existing clients when trying ti import data. If a client is found with the same client_id, instead of creating a nee client, the existing client will be updated and reports will be imported to this client insteadof creating a new client.

NOTE: this check is based on the client_id not the client name. Therefore, you will only be able to update client information from the same instance it was exported from