add option to Reports workflow to change how reports are imported
added second method to import reports that doesn't require selecting an existing client to import to. This method involves checking the client name on the saved ptrac and checking if there is an existing client with the same in the Plextrac instance. If so, the report will be imported under this client. If not, a new client will be created with just the client name in the ptrac file and the report will be imported under this new client.